Move those extend area into the composer namespace. Also, rename the folder.
