+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/BottomEditor/EditRowExtend.cs
@@ -10,7 +10,7 @@ using osu.Framework.Graphics.Shapes;
 using osu.Framework.Input.Events;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Extends
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.BottomEditor
 {
     public abstract class EditRowExtend : VisibilityContainer
     {
