+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/BottomEditor/Notes/NoteEditorHitObjectBlueprint.cs
@@ -22,7 +22,7 @@ using osu.Game.Rulesets.Karaoke.UI.Position;
 using osu.Game.Rulesets.UI.Scrolling;
 using osuTK;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Extends.Notes
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.BottomEditor.Notes
 {
     /// <summary>
     /// Copy from <see cref="NoteSelectionBlueprint"/>
