+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Compose/BottomEditor/TimeTags/TimeTagRowExtend.cs
@@ -7,13 +7,13 @@ using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Game.Rulesets.Karaoke.Objects;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Extends.Notes
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Compose.BottomEditor.TimeTags
 {
-    public class NoteRowExtend : EditRowExtend
+    public class TimeTagRowExtend : EditRowExtend
     {
-        public override float ContentHeight => 180;
+        public override float ContentHeight => 100;
 
-        public NoteRowExtend(Lyric lyric)
+        public TimeTagRowExtend(Lyric lyric)
             : base(lyric)
         {
         }
@@ -26,10 +26,10 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.LyricList.Rows.Extends.Notes
 
         protected override Drawable CreateContent(Lyric lyric)
         {
-            return new NoteEditor(lyric)
+            return new TimeTagEditor(lyric)
             {
                 RelativeSizeAxes = Axes.X,
-                Height = 150,
+                Height = 100,
             };
         }
     }
